# Added functional programming version of go