Made a version of the go interpreter that matches the Scala implementation instead of the standard struct based system.

Performance is better as well :)