@kostya 2016-12-21
hm seems this is slow for bench.b, 5.36s -> 5.57s, for mandel.b it faster 100s -> 70s, but for gccgo it also slower.