@lologarithm 2016-12-21
That seems strange it would be slower for bench.b but faster for mandel.b. 

I didn't test with gccgo locally (only go 1.7.3 on ubuntu 16.04).

I can do some more benching and see if I can figure out why that would be. Could you try a few times and see if that time was just a fluke?