@lologarithm 2016-12-21
mandel.b
Go
75.26s, 2.7Mb
Functional Go
48.05s, 1.7Mb

For bench.b I see
Go
4.34s, 1.4Mb
Functional Go
3.10s, 1.4Mb


I ran the tests about 10 times and didn't see much divergence from what I pasted here.